+++ packages/cmf/src/sagas/collection.js
@@ -1,5 +1,6 @@
 import { call, select } from 'redux-saga/effects';
 import { delay } from 'redux-saga';
+import selectors from '../selectors';
 
 /**
  * this saga ends when the collection is available
@@ -9,7 +10,7 @@ import { delay } from 'redux-saga';
 export function* waitFor(id, interval = 10) {
 	// eslint-disable-next-line no-constant-condition
 	while (true) {
-		const collection = yield select(state => state.cmf.collections.get(id));
+		const collection = yield select(selectors.collections.get);
 		if (collection !== undefined) {
 			break;
 		}
