+++ packages/containers/src/DeleteResource/deleteResource.sagas.js
@@ -1,7 +1,6 @@
 import invariant from 'invariant';
 import { take, put, race, call, select } from 'redux-saga/effects';
-import { actions } from '@talend/react-cmf';
-import { findCollectionPathListItem } from '@talend/react-cmf/lib/selectors';
+import { api, actions } from '@talend/react-cmf';
 import deleteResourceConst from './deleteResource.constants';
 
 /**
@@ -56,7 +55,7 @@ got ${resourcePath}`,
 export function* deleteResourceValidate(uri, resourceType, itemId, resourcePath) {
 	yield take(deleteResourceConst.DIALOG_BOX_DELETE_RESOURCE_OK);
 	const resourceLocator = getResourceLocator(resourceType, resourcePath);
-	const resource = yield select(findCollectionPathListItem, resourceLocator, itemId);
+	const resource = yield select(api.selectors.collections.findListItem, resourceLocator, itemId);
 	if (resource) {
 		yield put(
 			buildHttpDelete(
@@ -104,7 +103,8 @@ export default function deleteResource({
 				}),
 			});
 		} catch (error) {
-			invariant(`DeleteResource race failed :${error}`);
+			console.error(error);
+			invariant(process.env.NODE_ENV !== 'production', `DeleteResource race failed :${error}`);
 		} finally {
 			yield put({
 				type: deleteResourceConst.DIALOG_BOX_DELETE_RESOURCE_CLOSE,
